refactor, test: 기존의 JuiceMaker 내에 존재하던 make(of:), validateStock(of:), useStock(of:) 로직을 JuiceManufacturer 타입으로 분리
- JuiceMaker 내에 stock(of:) 로직이 추가됨에 따라, 기존의 make(of:), validateStock(of:), useStock(of:) 함수들과는 역할이 다름을 인지
- 위의 세 메서드를 주스를 만드는 역할만을 담당하는 타입으로 생성하게 됨.
- 초깃값을 동일하게 바인딩해주어야 함.
- 타입으로 분리함에 따라, make(of:) 실행을 성공했을 때, FruitStore 타입의 재고를 반환하도록 함.
- 재고를 현재 JuiceMaker의 재고에 업데이트 해주도록 함.
- 기존 JuiceMaker의 역할을 JuiceManufacturer으로 이전시킴에 따라 테스트 코드 또한 이동 및 수정.
